+++ tests/pr_test.go
@@ -13,7 +13,7 @@ import (
 
 const completeExampleTerraformDir = "examples/complete"
 const quickstartSolutionTerraformDir = "solutions/quickstart"
-const mirroringExampleTerraformDir = "examples/mirroring"
+const fsCloudTerraformDir = "examples/fscloud"
 
 // Use existing group for tests
 const resourceGroup = "geretain-test-event-streams"
@@ -82,10 +82,32 @@ func TestRunQuickstartSolution(t *testing.T) {
 	assert.NotNil(t, output, "Expected some output")
 }
 
-func TestRunMirroringExample(t *testing.T) {
+func TestRunFSCloudExample(t *testing.T) {
 	t.Parallel()
-
-	options := setupOptions(t, "es-mirror", mirroringExampleTerraformDir)
+	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
+		Testing:      t,
+		TerraformDir: fsCloudTerraformDir,
+		Prefix:       "es-fscloud",
+		/*
+		 Comment out the 'ResourceGroup' input to force this tests to create a unique resource group to ensure tests do
+		 not clash. This is due to the fact that an auth policy may already exist in this resource group since we are
+		 re-using a permanent HPCS instance. By using a new resource group, the auth policy will not already exist
+		 since this module scopes auth policies by resource group.
+		*/
+		//ResourceGroup:      resourceGroup,
+		BestRegionYAMLPath: regionSelectionPath,
+		TerraformVars: map[string]interface{}{
+			"existing_kms_instance_guid": permanentResources["hpcs_south"],
+			"kms_key_crn":                permanentResources["hpcs_south_root_key_crn"],
+			"mirroring_enabled":          true,
+			"mirroring_topic_patterns":   []string{"topic-1", "topic-2"},
+			"mirroring": map[string]interface{}{
+				"source_crn":   permanentResources["event_streams_us_south_crn"],
+				"source_alias": "source-alias",
+				"target_alias": "target-alias",
+			},
+		},
+	})
 
 	output, err := options.RunTestConsistency()
 	assert.Nil(t, err, "This should not have errored")
